add update list

+++ app/crud/todo_item.py
@@ -1,10 +1,13 @@
 from typing import Sequence
 
+from fastapi import HTTPException, status
 from sqlalchemy import Select
 from sqlalchemy.orm import selectinload
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from core.models import ToDoItem, ToDoList
+
+from core.models import ToDoItem, ToDoList, User
 from core.schemas.todo import ToDoItemUpdate
 from core.schemas.todo import ToDoItemCreate
 
@@ -20,11 +23,12 @@ async def get_all_todo_item(
 async def get_all_item_list(
         list_id: ToDoList,
         session: AsyncSession,
+        user: User
 ) -> Sequence[ToDoItem]:
     stmt = (
         Select(ToDoItem)
         .options(selectinload(ToDoItem.todo_list))
-        .where(ToDoItem.list_id == list_id.id)
+        .where((ToDoItem.list_id == list_id.id) & (user.id == list_id.user))
         .order_by(ToDoItem.id)
     )
     result = await session.scalars(stmt)
@@ -43,8 +47,14 @@ async def create_todo_item(
         todo_item: ToDoItemCreate,
 ) -> ToDoItem:
     todo_item = ToDoItem(**todo_item.model_dump())
-    session.add(todo_item)
-    await session.commit()
+    try:
+        session.add(todo_item)
+        await session.commit()
+    except IntegrityError:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="List missing at user"
+        )
     return todo_item
 
 
@@ -52,17 +62,31 @@ async def update_todo_item(
         item: ToDoItem,
         item_update: ToDoItemUpdate,
         session: AsyncSession,
+        user: User,
 ) -> ToDoItem:
-    for title, completed in item_update.model_dump(exclude_unset=True).items():
-        setattr(item, title, completed)
-
-    await session.commit()
-    return item
+    await session.get(User, user.id)
+    if result is not None:
+        item.title = item_update.title
+        item.completed = item_update.completed
+        await session.commit()
+        return item
+    raise HTTPException(
+        status_code=status.HTTP_404_NOT_FOUND,
+        detail="Item mission at user"
+    )
 
 
 async def del_todo_item(
         todo_item: ToDoItem,
-        session: AsyncSession
+        session: AsyncSession,
+        user: User,
 ) -> None:
+    if todo_item.todo_list.user != user.id:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Item missing at user"
+        )
+
     await session.delete(todo_item)
     await session.commit()
+
