+++ app/crud/todo_list.py
@@ -1,5 +1,6 @@
 from typing import Sequence
 
+from fastapi import HTTPException, status
 from sqlalchemy import Select
 from sqlalchemy.orm import selectinload
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -23,10 +24,10 @@ async def get_all_todo_list(
 
 
 async def get_todo_list(
-    item_id: int,
+    list_id: int,
     session: AsyncSession,
 ) -> ToDoList | None:
-    return await session.get(ToDoList, item_id)
+    return await session.get(ToDoList, list_id)
 
 
 async def create_todo_list(
@@ -40,32 +41,53 @@ async def create_todo_list(
     return todo_list
 
 
-async def del_todo_list(todo_list: ToDoList, session: AsyncSession) -> None:
-    full_item = await get_all_item_list(list_id=todo_list, session=session)
+async def del_todo_list(
+        todo_list: ToDoList,
+        session: AsyncSession,
+        user: User,
+) -> None:
+    full_item = await get_all_item_list(list_id=todo_list, session=session, user=user)
     if full_item is not None:
         for item in full_item:
-            await del_todo_item(todo_item=item, session=session)
+            await del_todo_item(todo_item=item, session=session, user=user)
 
-    await session.delete(todo_list)
-    await session.commit()
+    if todo_list.user == user.id:
+        await session.delete(todo_list)
+        await session.commit()
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="List missing at user"
+        )
 
 
 async def update_title_list(
     list_at: ToDoList,
     list_update: ToDoListUpdate,
+    user: User,
     session: AsyncSession,
 ) -> ToDoList:
-    new_title = list_update.model_dump().get("title")
-    list_at.title = new_title
+    if list_at.user == user.id:
+        new_title = list_update.model_dump().get("title")
+        list_at.title = new_title
+        list_at.user = user.id
 
-    await session.commit()
+        await session.commit()
 
-    return list_at
+        return list_at
+
+    raise HTTPException(
+        status_code=status.HTTP_404_NOT_FOUND,
+        detail="List missing at user"
+    )
 
 
-async def full_todo(session: AsyncSession):
+async def full_todo(
+        session: AsyncSession,
+        user: User
+):
     stmt = (
-        Select(ToDoList).options(selectinload(ToDoList.todo_item)).order_by(ToDoList.id)
+        Select(ToDoList).options(selectinload(ToDoList.todo_item)).where(ToDoList.user == user.id).order_by(ToDoList.id)
     )
     result = await session.scalars(stmt)
     return result.all()
